+++ priv/repo/migrations/20260106154322_remove_nano_old_bootloader_firmware.exs
@@ -1,0 +1,10 @@
+defmodule Trenino.Repo.Migrations.RemoveNanoOldBootloaderFirmware do
+  use Ecto.Migration
+
+  def change do
+    execute(
+      "DELETE FROM firmware_files WHERE board_type = 'nano_old_bootloader'",
+      "SELECT 1"
+    )
+  end
+end